refactor!:Source paths rename to match working repo
