--- src/main/java/com/github/tkutcher/jgrade/OutputFormatter.java
@@ -1,13 +1,0 @@
-package com.github.tkutcher.jgrade;
-
-/**
- * A formatter that formats data of a {@link Grader} to produce output.
- */
-public interface OutputFormatter {
-
-    /**
-     * Get the formatted output of the grader.
-     * @param grader The grader observing.
-     */
-    String format(Grader grader);
-}
