+++ src/test/java/com/github/dscpsyl/jgrade/gradedtest/AllGraderTests.java
@@ -1,0 +1,11 @@
+package com.github.dscpsyl.jgrade.gradedtest;
+
+import org.junit.runner.RunWith;
+import org.junit.runners.Suite;
+
+@RunWith(Suite.class)
+@Suite.SuiteClasses({
+        GradedTestListenerTest.class,
+        GradedTestResultTest.class,
+})
+public class AllGraderTests { }
